+++ app/Http/Controllers/Manager/OptionsController.php
@@ -21,7 +21,10 @@ declare(strict_types = 1);
 
 namespace Adshares\Adserver\Http\Controllers\Manager;
 
+use Adshares\Adserver\Client\Mapper\AbstractFilterMapper;
 use Adshares\Adserver\Http\Controller;
+use Adshares\Adserver\Http\Requests\TargetingReachRequest;
+use Adshares\Adserver\Services\Advertiser\TargetingReachComputer;
 use Adshares\Adserver\ViewModel\OptionsSelector;
 use Adshares\Common\Application\Service\ConfigurationRepository;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -50,6 +53,18 @@ class OptionsController extends Controller
     public function targeting(): JsonResponse
     {
         return self::json(new OptionsSelector($this->optionsRepository->fetchTargetingOptions()));
+    }
+
+    public function targetingReach(TargetingReachRequest $request): JsonResponse
+    {
+        $targeting = $request->toArray()['targeting'];
+
+        $requires = AbstractFilterMapper::generateNestedStructure($targeting['requires']);
+        $excludes = AbstractFilterMapper::generateNestedStructure($targeting['excludes']);
+
+        $targetingReach = (new TargetingReachComputer())->compute($requires, $excludes);
+
+        return self::json($targetingReach->toArray());
     }
 
     public function filtering(): JsonResponse