+++ routes/manager.php
@@ -83,11 +83,13 @@ Route::middleware([Kernel::USER_ACCESS, Kernel::JSON_API])->group(function () {
 
     Route::get('options/campaigns', [OptionsController::class, 'campaigns']);
     Route::get('options/campaigns/targeting', [OptionsController::class, 'targeting']);
+    Route::post('options/campaigns/targeting-reach', [OptionsController::class, 'targetingReach']);
     Route::get('options/sites/filtering', [OptionsController::class, 'filtering']);
     Route::get('options/sites/languages', [OptionsController::class, 'languages']);
     Route::get('options/sites/zones', [OptionsController::class, 'zones']);
 
 // Routes for Withdraw/Deposit
+    Route::get('withdrawal-info', [WalletController::class, 'withdrawalInfo']);
     Route::post('calculate-withdrawal', [WalletController::class, 'calculateWithdrawal']);
     Route::post('wallet/withdraw', [WalletController::class, 'withdraw']);
     Route::get('deposit-info', [WalletController::class, 'depositInfo']);